Toggle private and public bookmarks

+++ dre_django/bookmarksapp/views.py
@@ -47,6 +47,9 @@ def toggle_bookmark( request, ctype_id, object_id ):
     bookmark = get_bookmark( user, obj )
 
     if bookmark: 
+        # Only the owner of the bookmark can delete it
+        if bookmark.user != request.user:
+            raise Http404
         bookmark.delete()
     else:
         bookmark = add_bookmark( user, obj )
@@ -55,4 +58,33 @@ def toggle_bookmark( request, ctype_id, object_id ):
         return redirect(redirect_to)
     else:
         return HttpResponse('<h1>Bookmark toggled</h1>') 
+
  
+@login_required
+def toggle_public( request, ctype_id, object_id ):
+    user = request.user
+    content_type = ContentType.objects.get(id=ctype_id)
+
+    try:
+        obj = content_type.get_object_for_this_type(id=object_id)
+    except ObjectDoesNotExist:
+        raise Http404
+
+    redirect_to = request.REQUEST.get('next', '')
+
+    bookmark = get_bookmark( user, obj )
+
+    if not bookmark: 
+        raise Http404
+    if bookmark.user != request.user:
+        # Only the owner of the bookmark can change its status 
+        raise Http404
+
+    # Toggle the public flag
+    bookmark.public = not bookmark.public
+    bookmark.save()
+
+    if redirect_to:
+        return redirect(redirect_to)
+    else:
+        return HttpResponse('<h1>Bookmark status toggled</h1>')
